refactor: extract test coverage symbol helpers
